defectdojo_apiv2: fixes to set_finding (#59)
* defectdojo_apiv2: update parameters of set_findings

DefectDojo requires fields title, description, severity and
numerical_severity to be present and not null. Added missing
numerical security so that in can be included, and relaxed
other fields so that they can be omitted.

https://github.com/DefectDojo/django-DefectDojo/blob/master/dojo/models.py#L1607
https://github.com/DefectDojo/django-DefectDojo/blob/master/dojo/api_v2/serializers.py#L928

Unsure if there are existing users with ordered arguments.
Without changing parameter ordering, added checks for mandatory
parameters.

* defectdojo_apiv2: bugfixes to set_finding

Code adds only given parameters to API request.

For two fields, the if-test misses False values.
Instead of a simple if, boolean values are compared
against None explicitly.

With this, active and verified boolean fields can
also be set as False.